refactor(General): :recycle: Unify queries enums with section enums
