--- src/pages/ProcessGroups/services/services.enum.ts
@@ -1,5 +1,0 @@
-export enum QueriesProcessGroups {
-  GetProcessGroups = 'get-process-groups-query',
-  GetProcessGroup = 'get-process-groups-query',
-  GetProcessesByProcessGroup = 'get-processes-by-process-group-query'
-}
