+++ src/pages/Sites/Sites.enum.ts
@@ -2,6 +2,17 @@ export enum SitesRoutesPaths {
   Sites = '/sites'
 }
 
+export enum QueriesSites {
+  GetSites = 'get-sites-query',
+  GetSite = 'get-site-query',
+  GetSitesPairs = 'get-sites-pairs-query',
+  GetHostsBySiteId = 'get-hosts-by-site-id-query',
+  GetLinksBySiteId = 'get-links-by-site-id-query',
+  GetRouters = 'get-routers-query',
+  GetLinks = 'get-links-query',
+  GetProcessesBySiteId = 'get-processes-by-site-id-query'
+}
+
 export enum Labels {
   Section = 'Sites',
   Description = 'A site is a location where components of your application are running and they are linked together to form a network',
