--- src/pages/Sites/services/services.enum.ts
@@ -1,10 +1,0 @@
-export enum QueriesSites {
-  GetSites = 'get-sites-query',
-  GetSite = 'get-site-query',
-  GetSitesPairs = 'get-sites-pairs-query',
-  GetHostsBySiteId = 'get-hosts-by-site-id-query',
-  GetLinksBySiteId = 'get-links-by-site-id-query',
-  GetRouters = 'get-routers-query',
-  GetLinks = 'get-links-query',
-  GetProcessesBySiteId = 'get-processes-by-site-id-query'
-}
